refactor: combine todoMap and todoOrder into a single state and extract Entry component
- Merged `todoMap` and `todoOrder` into a single state object for better state management.
- Extracted `Entry` component into its own file to improve code modularity and readability.
